refactor(tx): split command.py God Object and relocate QoS
- Relocated the `Qos` class to `protocol/qos.py` alongside `QosManager` to unify QoS responsibilities.
- Split the monolithic `command.py` into a modular `command/` package utilizing domain-specific mixins (`dhw`, `hvac`, `system`, `zones`).
- Enforced strict `mypy` compliance across the new module boundaries using generic `TypeVar` binding.
- Ensured zero breaking changes to the upstream `Command` API footprint.
- Preserved all original inline comments, TODOs, and technical debt anchors verbatim.

This completes Phase 1 of the architectural restructuring required to address Issue #591.
